Move typeck into its own crate.
