+++ mk/crates.mk
@@ -53,7 +53,7 @@ TARGET_CRATES := libc std flate arena term \
                  serialize getopts collections test time rand \
                  log regex graphviz core rbml alloc rustrt \
                  unicode
-HOST_CRATES := syntax rustc rustc_trans rustdoc regex_macros fmt_macros \
+HOST_CRATES := syntax rustc rustc_typeck rustc_trans rustdoc regex_macros fmt_macros \
 	       rustc_llvm rustc_back
 CRATES := $(TARGET_CRATES) $(HOST_CRATES)
 TOOLS := compiletest rustdoc rustc
@@ -67,7 +67,9 @@ DEPS_std := core libc rand alloc collections rustrt unicode \
 	native:rust_builtin native:backtrace
 DEPS_graphviz := std
 DEPS_syntax := std term serialize log fmt_macros arena libc
-DEPS_rustc_trans := rustc rustc_back rustc_llvm libc
+DEPS_rustc_trans := arena flate getopts graphviz libc rustc rustc_back \
+                    rustc_typeck log syntax serialize rustc_llvm
+DEPS_rustc_typeck := rustc syntax
 DEPS_rustc := syntax flate arena serialize getopts rbml \
               time log graphviz rustc_llvm rustc_back
 DEPS_rustc_llvm := native:rustllvm libc std
@@ -110,8 +112,11 @@ ONLY_RLIB_unicode := 1
 # You should not need to edit below this line
 ################################################################################
 
-DOC_CRATES := $(filter-out rustc, $(filter-out rustc_trans, $(filter-out syntax, $(CRATES))))
-COMPILER_DOC_CRATES := rustc rustc_trans syntax
+DOC_CRATES := $(filter-out rustc, \
+              $(filter-out rustc_trans, \
+              $(filter-out rustc_typeck, \
+              $(filter-out syntax, $(CRATES)))))
+COMPILER_DOC_CRATES := rustc rustc_trans rustc_typeck syntax
 
 # This macro creates some simple definitions for each crate being built, just
 # some munging of all of the parameters above.
