+++ src/librustc_typeck/check/regionmanip.rs
