+++ src/librustc_typeck/check/writeback.rs
@@ -13,17 +13,17 @@
 // substitutions.
 use self::ResolveReason::*;
 
+use astconv::AstConv;
+use check::FnCtxt;
 use middle::def;
 use middle::pat_util;
 use middle::ty::{mod, Ty, MethodCall, MethodCallee};
 use middle::ty_fold::{TypeFolder,TypeFoldable};
-use typeck::astconv::AstConv;
-use typeck::check::FnCtxt;
 use middle::infer::{force_all, resolve_all, resolve_region};
 use middle::infer::resolve_type;
 use middle::infer;
-use typeck::write_substs_to_tcx;
-use typeck::write_ty_to_tcx;
+use write_substs_to_tcx;
+use write_ty_to_tcx;
 use util::ppaux::Repr;
 
 use std::cell::Cell;
