+++ src/librustc_typeck/diagnostics.rs
@@ -145,5 +145,7 @@ register_diagnostics!(
     E0166,
     E0167,
     E0168,
-    E0169
+    E0169,
+    E0171,
+    E0172
 )
